aks-workshop/py/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
pyyaml==5.3.1             # via -r requirements.in
